Prescripto/backend/chatbot/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
requests==2.31.0
werkzeug==2.0.3 